IOS-3788: Split scroll view connecting and scrolling into separate entities, naming improvements
